resume_screening_app9/requirements.txt
langchain
streamlit
tiktoken
unstructured
pinecone-client
pypdf
sentence-transformers
langchain-google-genai
uuid
langchain-community